Splitting aligner.h/cpp files
