+++ include/vrv/attcomparison.h
@@ -8,7 +8,7 @@
 #ifndef __VRV_ATT_COMPARISON_H__
 #define __VRV_ATT_COMPARISON_H__
 
-#include "aligner.h"
+#include "horizontalaligner.h"
 #include "artic.h"
 #include "atts_shared.h"
 #include "durationinterface.h"
